move the register.html to the views folder
