--- public/signing.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html>
- <head>
-  <title>Lotus Software</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-  <link href="style.css" rel="stylesheet">
-</head>
- <body>
-   <form action="action_page.php">
-     <div class="container">
-       <h1>Register</h1>
-       <p>Please fill in this form to create an account!</p>
-
-       <label for="username"><b>Username</b></label>
-       <input type="text" placeholder="Enter your Username" name="username" required>
-
-       <label for="email"><b>Email</b></label>
-       <input type="text" placeholder="Enter Email" name="email" required>
-
-       <label for="psw"><b>Password</b></label>
-       <input type="password" placeholder="Enter Password" name="psw" required>
-
-       <label for="confirm-psw"><b>Confirm Password</b></label>
-       <input type="password" placeholder="Confirm Password" name="confirmPsw" required>
-       <hr>
-
-       <button type="submit" class="registerbtn">Sign up</button>
-     </div>
-
-         <button type="submit" class="registerbtn" formaction="/signbtn">Sign up</button>
-       </div>
-
-       <div class="container signing">
-         <p>Already have an account?</p>
-         <a href="./login">Sign in</a>
-       </div>
-     </form>
-
-<!-- </script> -->
-</body>
- </html>
