+++ src/database/db_build.sql
@@ -1,17 +1,23 @@
 BEGIN;
 
 DROP TABLE IF EXISTS users;
+DROP TABLE IF EXISTS info;
+
 
 CREATE TABLE users (
   user_id SERIAL PRIMARY KEY,
   username VARCHAR(30) NOT NULL,
+  email VARCHAR(100) NOT NULL,
+  password VARCHAR(100) NOT NULL
+);
+
+CREATE TABLE info (
   first_name VARCHAR(30) NOT NULL,
   last_name VARCHAR(30) NOT NULL,
-  email VARCHAR(100) NOT NULL,
-  password VARCHAR(100) NOT NULL,
   age VARCHAR(2) NOT NULL,
   gender VARCHAR(30) NOT NULL,
-  high_school_deploma VARCHAR(3) NOT NULL
+  high_learning VARCHAR(3) NOT NULL,
+  email VARCHAR(100) NOT NULL
 );
 
 COMMIT;
